assistant2: Update to match latest designs
